Refactor OAuth integration: Move sources, errors, and protocols to utils package
