+++ integrations/oauth/pyproject.toml
@@ -70,7 +70,8 @@ integration = 'pytest -m "integration" {args:tests}'
 all = 'pytest {args:tests}'
 unit-cov-retry = 'pytest --cov=haystack_integrations --reruns 3 --reruns-delay 30 -x -m "not integration" {args:tests}'
 integration-cov-append-retry = 'pytest --cov=haystack_integrations --cov-append --reruns 3 --reruns-delay 30 -x -m "integration" {args:tests}'
-types = "mypy -p haystack_integrations.components.connectors.oauth {args}"
+types = """mypy -p haystack_integrations.components.connectors.oauth \
+-p haystack_integrations.utils.oauth {args}"""
 
 [tool.mypy]
 install_types = true
