+++ integrations/oauth/src/haystack_integrations/components/connectors/oauth/__init__.py
@@ -1,24 +1,6 @@
 # SPDX-FileCopyrightText: 2022-present deepset GmbH <info@deepset.ai>
 #
 # SPDX-License-Identifier: Apache-2.0
-from .errors import OAuthConfigError, OAuthError, TokenRefreshError
 from .resolver import OAuthResolver
-from .sources import (
-    RefreshTokenSource,
-    StaticTokenSource,
-    SubjectTokenSource,
-    TokenExchangeSource,
-    TokenSource,
-)
 
-__all__ = [
-    "OAuthConfigError",
-    "OAuthError",
-    "OAuthResolver",
-    "RefreshTokenSource",
-    "StaticTokenSource",
-    "SubjectTokenSource",
-    "TokenExchangeSource",
-    "TokenRefreshError",
-    "TokenSource",
-]
+__all__ = ["OAuthResolver"]
