+++ integrations/oauth/src/haystack_integrations/utils/oauth/__init__.py
@@ -1,0 +1,17 @@
+# SPDX-FileCopyrightText: 2022-present deepset GmbH <info@deepset.ai>
+#
+# SPDX-License-Identifier: Apache-2.0
+from .errors import OAuthConfigError, OAuthError, TokenRefreshError
+from .protocols import SubjectTokenSource, TokenSource
+from .sources import RefreshTokenSource, StaticTokenSource, TokenExchangeSource
+
+__all__ = [
+    "OAuthConfigError",
+    "OAuthError",
+    "RefreshTokenSource",
+    "StaticTokenSource",
+    "SubjectTokenSource",
+    "TokenExchangeSource",
+    "TokenRefreshError",
+    "TokenSource",
+]
