+++ integrations/oauth/src/haystack_integrations/utils/oauth/errors.py
